py/requirements.txt
absl-py==1.1.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
black==22.3.0
cachetools==5.2.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
cssselect2==0.6.0
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.0
fastapi==0.75.2
filelock==3.6.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
html5lib==1.1
httptools==0.4.0
huggingface-hub==0.7.0
identify==2.5.0
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
jarowinkler==1.0.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.4.2
mplcursors==0.5.1
mypy==0.950
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
orjson==3.6.8
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pre-commit==2.19.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.9
pypdfium2==2.0.0
pyphen==0.12.0
python-dateutil==2.8.2
python-doctr==0.5.1
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
rapidfuzz==2.0.11
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-addons==0.17.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typeguard==2.13.3
typing_extensions==4.2.0
ujson==5.2.0
Unidecode==1.3.4
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
watchgod==0.8.2
WeasyPrint==52.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.2
wrapt==1.14.1
